Moved jsonpath parsing logic out of template logic.
